requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.29.0
redis==5.0.3
pydantic==2.10.6
pydantic-settings==2.8.1
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.27.0
pytest-cov==5.0.0
fakeredis[lua]==2.21.3
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
